Move function sonify_point to spectrograms.py to fix problem with circular imports
